feat: Implement JS transform caching and expression evaluation
